Move visualization panel to visualizations
