+++ client/src/components/Visualizations/VisualizationPanel.vue
